+++ dkron/rpc.go
@@ -2,11 +2,14 @@ package dkron
 
 import (
 	"errors"
+	"fmt"
 	"net"
 	"net/http"
 	"net/rpc"
+	"time"
 
 	"github.com/Sirupsen/logrus"
+	metrics "github.com/armon/go-metrics"
 	"github.com/docker/libkv/store"
 	"github.com/hashicorp/serf/serf"
 )
@@ -20,6 +23,7 @@ type RPCServer struct {
 }
 
 func (rpcs *RPCServer) GetJob(jobName string, job *Job) error {
+	defer metrics.MeasureSince([]string{"rpc", "get_job"}, time.Now())
 	log.WithFields(logrus.Fields{
 		"job": jobName,
 	}).Debug("rpc: Received GetJob")
@@ -37,6 +41,7 @@ func (rpcs *RPCServer) GetJob(jobName string, job *Job) error {
 }
 
 func (rpcs *RPCServer) ExecutionDone(execution Execution, reply *serf.NodeResponse) error {
+	defer metrics.MeasureSince([]string{"rpc", "execution_done"}, time.Now())
 	log.WithFields(logrus.Fields{
 		"group": execution.Group,
 		"job":   execution.JobName,
@@ -135,8 +140,13 @@ func listenRPC(a *AgentCommand) {
 		agent: a,
 	}
 
+	bindIp, err := a.GetBindIP()
+	if err != nil {
+		log.Fatal("rpc:", err)
+	}
+	RPCAddr := fmt.Sprintf("%s:%d", bindIp, a.config.RPCPort)
 	log.WithFields(logrus.Fields{
-		"rpc_addr": a.getRPCAddr(),
+		"rpc_addr": RPCAddr,
 	}).Debug("rpc: Registering RPC server")
 
 	rpc.Register(r)
@@ -157,7 +167,7 @@ func listenRPC(a *AgentCommand) {
 	// workaround
 	http.DefaultServeMux = oldMux
 
-	l, e := net.Listen("tcp", a.getRPCAddr())
+	l, e := net.Listen("tcp", RPCAddr)
 	if e != nil {
 		log.Fatal("rpc:", e)
 	}
@@ -170,6 +180,7 @@ type RPCClient struct {
 }
 
 func (rpcc *RPCClient) callExecutionDone(execution *Execution) error {
+	defer metrics.MeasureSince([]string{"rpc", "call_execution_done"}, time.Now())
 	client, err := rpc.DialHTTP("tcp", rpcc.ServerAddr)
 	if err != nil {
 		log.WithFields(logrus.Fields{
@@ -195,6 +206,7 @@ func (rpcc *RPCClient) callExecutionDone(execution *Execution) error {
 }
 
 func (rpcc *RPCClient) GetJob(jobName string) (*Job, error) {
+	defer metrics.MeasureSince([]string{"rpc", "call_get_job"}, time.Now())
 	client, err := rpc.DialHTTP("tcp", rpcc.ServerAddr)
 	if err != nil {
 		log.WithFields(logrus.Fields{